refactor: terminal / run command
